ChessClockAPI/requirements.txt
channels
channels-redis
Django
django-colorfield
dj_database_url
psycopg2
pyshorteners
djangorestframework
django-filter
grpcio
whitenoise

